⚡️ Modify : 코드 리팩토링 및 query, mutation 분리
- 코드 리팩토링 및 query, mutation 분리
