+++ src/components/educator/item/inquiryItem/DeleteItemModal.tsx
@@ -1,54 +1,24 @@
+import Spinner from "@/components/common/spinner/Spinner";
+import useDeleteItems from "@/hooks/mutations/item/useDeleteItems";
 import { deleteItemsModalState } from "@/states/modalState/confirmModalState";
 import { selectedItemState } from "@/states/selectedState/selectedItemState";
 import { useEffect, useRef } from "react";
-import { useRecoilState, useResetRecoilState } from "recoil";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import useAxios from "@/hooks/useAxios";
-import { useParams } from "react-router-dom";
-import Spinner from "@/components/common/spinner/Spinner";
+import { useRecoilState, useRecoilValue, useResetRecoilState } from "recoil";
 
 const DeleteItemModal = () => {
-	const [selectedItems, setSelectedItems] = useRecoilState(selectedItemState);
+	const selectedItems = useRecoilValue(selectedItemState);
 	const [isOpenModal, setIsOpenModal] = useRecoilState(deleteItemsModalState);
 	const resetIsOpenModal = useResetRecoilState(deleteItemsModalState);
-	const { axiosData, isFetchLoading } = useAxios();
-	const { channelId } = useParams();
 	const modalRef = useRef<HTMLDivElement>(null);
-	const queryClient = useQueryClient();
+	const { mutate, isPending } = useDeleteItems();
 
 	const handleClickCancelBtn = () => {
 		setIsOpenModal(false);
 	};
 
 	const handleClickDeleteBtn = () => {
-		deleteItemsMutation.mutate();
+		mutate();
 	};
-
-	const deleteItemsData = async () => {
-		const response = await axiosData("useToken", {
-			method: "DELETE",
-			url: `/teachers/api/v1/channels/${channelId}/items`,
-			data: {
-				itemIds: selectedItems,
-			},
-		});
-
-		if (response) {
-			const status = response.status;
-
-			if (status === 204) {
-				alert("삭제가 완료되었습니다.");
-				queryClient.invalidateQueries({ queryKey: ["items", channelId] });
-				setSelectedItems([]);
-				setIsOpenModal(false);
-			}
-		}
-	};
-
-	const deleteItemsMutation = useMutation({
-		mutationKey: ["deleteItems"],
-		mutationFn: deleteItemsData,
-	});
 
 	useEffect(() => {
 		const handleOutSideClick = (e: Event) => {
@@ -74,7 +44,7 @@ const DeleteItemModal = () => {
 
 	return (
 		<>
-			{isFetchLoading ? (
+			{isPending ? (
 				<Spinner />
 			) : (
 				<div
