+++ src/components/tradeStock/tradeHome/TradeHomeContainer.tsx
@@ -1,45 +1,14 @@
 import Spinner from "@/components/common/spinner/Spinner";
-import useAxios from "@/hooks/useAxios";
+import useAllStockList from "@/hooks/queries/tradeStock/useAllStockList";
 import { useIntersectionObserver } from "@/hooks/useObserver";
-import { IStockHomeResponseData } from "@/interface/stock";
-import { userRoleState } from "@/states/userRoleState";
-import { useInfiniteQuery } from "@tanstack/react-query";
-import { useParams } from "react-router-dom";
-import { useRecoilValue } from "recoil";
 import DeleteRegisterButton from "./DeleteRegisterButton";
 import DeleteStocksModal from "./DeleteStocksModal";
 import StockList from "./StockList";
 import TradeHomeHeading from "./TradeHomeHeading";
 
-interface IInfinityQueryData {
-	readonly pageParams: number[];
-	readonly pages: IStockHomeResponseData[];
-}
-
 const TradeHomeContainer = () => {
-	const userRole = useRecoilValue(userRoleState);
-	const { channelId } = useParams();
-	const { axiosData } = useAxios();
-
-	const getTradeStocksData = async (pageParam: number) => {
-		const response = await axiosData("useToken", {
-			method: "GET",
-			url: `/${userRole}s/api/v1/channels/${channelId}/stocks?limit=15&offset=${pageParam}`,
-		});
-
-		const resultData = response?.data.data;
-		return resultData;
-	};
-
 	const { data, isFetching, isLoading, hasNextPage, fetchNextPage } =
-		useInfiniteQuery<IStockHomeResponseData, Error, IInfinityQueryData>({
-			queryKey: ["stocks", channelId],
-			queryFn: ({ pageParam }) => getTradeStocksData(pageParam as number),
-			initialPageParam: 0,
-			getNextPageParam: (lastPage) => {
-				return lastPage.next !== null ? lastPage.offset + 15 : undefined;
-			},
-		});
+		useAllStockList();
 
 	const observeTargetRef = useIntersectionObserver({
 		hasNextPage,
