+++ src/states/studentItem/studentSelectedItem.ts
@@ -1,7 +1,7 @@
-import { IStudentMyItem } from "@/interface/item";
+import { IExtendItem } from "@/interface/item";
 import { atom } from "recoil";
 
-export const studentSelectedItem = atom<IStudentMyItem | null>({
+export const studentSelectedItem = atom<IExtendItem | null>({
 	default: null,
 	key: "studentSelectedItem",
 });
